Wowza Begin Implementation

+++ api/models/User.js
@@ -1,5 +1,9 @@
 'use strict';
 
+/**
+ * Module dependencies.
+ * @private
+ */
 var connection = require('../database/mysql');
 
 /**
@@ -26,6 +30,11 @@ CREATE UNIQUE INDEX unique_email
 ON user(email);
 */
 
+/**
+ * Module class
+ * @public
+ */
+
 var User = function(model) {
 	this.id 		 = model.id;
 	this.username 	 = model.username;
@@ -40,7 +49,12 @@ var User = function(model) {
 	this.createdDate = model.createdDate;
 };
 
-/// Controller Methods
+/**
+ * Module methods.
+ * @public
+ */
+ 
+ // return all records in 'user' table
 User.getAll = (res) => {
 	connection.query(
 		"SELECT * FROM user ORDER BY id",
@@ -52,7 +66,21 @@ User.getAll = (res) => {
 		});
 };
 
+/**
+ * Return record from 'user' table with matching 'id'
+ *
+ * @param {int} id
+ * @param {function} res
+ *
+ * @return {function} 
+ * @public
+ */
 User.getById = (id, res) => {
+	
+	if (isNaN(id)) {
+		throw new TypeError(`Argument expected instance of integer, received instance of ${typeof id}.`);
+	}
+	
 	connection.query(
 		"SELECT * FROM user WHERE id = ? ORDER BY id",
 		id,
@@ -66,9 +94,26 @@ User.getById = (id, res) => {
 	);
 };
 
+/**
+ * Returns record from 'user' table with matching 'username'
+ *
+ * @param {string} username
+ * @param {function} res
+ *
+ * @public
+ */
 User.getByUsername = (username, res) => {
+	
+	if (typeof username !== 'string') {
+		throw new TypeError(`Argument expected instance of string, received instance of ${typeof username}.`);
+	}
+	
+	if (username.length === 0) {
+		throw new Error('Argument must not be empty.');
+	}
+	
 	connection.query(
-		"SELECT * FROM user WHERE username = ? ORDER BY username",
+		"SELECT * FROM user WHERE username = ? LIMIT 1",
 		username,
 		(error, data) => {
 			if (error) {
@@ -80,7 +125,23 @@ User.getByUsername = (username, res) => {
 	);
 };
 
+/**
+ * Returns record with matching username or email
+ *
+ * @param {string} input
+ * @param {function} res
+ * @public
+ */
 User.getByUsernameOrEmail = (input, res) => {
+	
+	if (typeof input !== 'string') {
+		throw new TypeError(`Argument expected instance of string, received instance of ${typeof input}.`);
+	}
+	
+	if (input.length === 0) {
+		throw new Error('Argument must not be empty.');
+	}
+	
 	connection.query(
 		"SELECT * FROM user WHERE username = ? OR email = ? LIMIT 1",
 		[input, input],
@@ -94,7 +155,19 @@ User.getByUsernameOrEmail = (input, res) => {
 	);
 };
 
+/**
+ * Creates new record in 'user' table
+ *
+ * @param {object} user
+ * @param {function} res
+ * @public
+ */
 User.create = (user, res) => {
+	
+	if (typeof user !== 'object') {
+		throw new TypeError(`Argument expected instance of object, received instance of ${typeof user}.`);
+	}
+	
 	connection.query(
 		"INSERT INTO user SET ?",
 		user,
@@ -107,6 +180,12 @@ User.create = (user, res) => {
 	);
 };
 
+/**
+ * Deletes all records in 'user' table
+ *
+ * @param {function} res
+ * @public
+ */
 User.deleteAll = (res) => {
 	connection.query(
 		"DELETE FROM user WHERE id > 0",
@@ -119,8 +198,19 @@ User.deleteAll = (res) => {
 	);
 };
 
-
+/**
+ * Deletes record in 'user' table with matching 'id'
+ * 
+ * @param {int} id
+ * @param {function} res
+ * @public
+ */
 User.deleteById = (id, res) => {
+	
+	if (isNaN(id)) {
+		throw new TypeError(`Argument expected an instance of integer, received instance of ${typeof id}.`);
+	}
+	
 	connection.query(
 		"DELETE FROM user WHERE id = ?",
 		id,
@@ -133,8 +223,28 @@ User.deleteById = (id, res) => {
 	);
 };
 
-// MARK: - PUT
+/**
+ * Updates record in 'user' table with 'id'
+ *
+ * @param {int} id
+ * @param {string} preferredName
+ * @param {function} res
+ * @public
+ */
 User.updatePreferredName = (id, preferredName, res) => {
+	
+	if (isNaN(id)) {
+		throw new TypeError(`Argument expected an instance of integer, received instance of ${typeof id}.`);
+	}
+	
+	if (typeof preferredName !== 'string') {
+		throw new TypeError(`Argument expected an instance of string, received instance of ${typeof preferredName}.`);
+	}
+	
+	if (preferredName.count === 0) {
+		throw new Error('Argument must not be empty.');
+	}
+		
 	connection.query(
 		"UPDATE user SET preferredName = ? WHERE id  = ?",
 		[preferredName, id],
@@ -147,10 +257,27 @@ User.updatePreferredName = (id, preferredName, res) => {
 	);
 };
 
-User.updateLastLogin = (id, lastLogin, res) => {
+/**
+ * Updates column 'lastLogin' in 'user' table with matching id
+ *
+ * @param {int} id
+ * @param {date} lastLoginDate
+ * @param {function} res
+ * @public
+ */
+User.updateLastLogin = (id, lastLoginDate, res) => {
+	
+	if (isNaN(id)) {
+		throw new TypeError(`Argument expected an instance of integer, received instance of ${typeof id}.`);
+	}
+	
+	if (typeof lastLoginDate !== 'date') {
+		throw new TypeError(`Argument expected an instance of date, received instance of ${typeof lastLogin}.`);
+	}
+	
 	connection.query(
 		"UPDATE user SET lastLogin = ? WHERE id = ?",
-		[lastLogin, id],
+		[lastLoginDate, id],
 		(error, data) => {
 			if (error)
 				res(error, null);
@@ -158,8 +285,39 @@ User.updateLastLogin = (id, lastLogin, res) => {
 				res(null, data);
 		}
 	);
-}
+};
 
+/**
+ * Updates column 'dateOfBirth' in 'user' table for record with matching id
+ *
+ * @param {int} id
+ * @param {string} dob
+ * @param {function} res
+ * @public
+ */
+ User.updateDateOfBirth = (id, dob, res) => {
+ 	
+ 	if (isNaN(id)) {
+ 		throw new TypeError(`Argument expected an instance of integer, received instance of ${typeof id}.`);
+ 	}
+ 	
+ 	if (Date.parse(dob)) {
+ 		throw new TypeError(`Argument is invalid date format.`);
+ 	}
+ 	
+ 	var dateOfBirth = new Date(dob);
+ 	
+ 	connection.query(
+ 		"UPDATE user SET dateOfBirth = ? WHERE id = ?",
+ 		[dateOfBirth, id],
+ 		(error, data) => {
+ 			if (error)
+ 				res(error, null);
+ 			else
+ 				res(null, data);
+ 		}
+ 	);
+ };
 
 
 module.exports = User;
